fix(microservices) updated models hash and user service
